requirements.txt
chumpy @ git+https://github.com/mattloper/chumpy
yacs
joblib
ffmpeg-python                 
munkres                       
numpy <= 1.23.0                     
nvidia-tensorrt==8.4.1.5             
onnx                          
onnx-simplifier               
onnxruntime                  
onnxsim                       
opencv-python                 
packaging                     
setuptools
torch                                     
torchaudio                    
torchvision 
detectron2
pandas        
pyyaml
tqdm
matplotlib
scikit-image
imageio[ffmpeg]
tensorboard
smplx
timm
scipy
lap
Cython
psutil
ultralytics
cython-bbox
progress
einops
mmcv==2.0.0rc4
xtcocotools>=1.8
loguru
gdown==4.6.0